Move jsonb column type from cost_model.go
