+++ internal/datasource/cardano_db_sync/models/treasury.go
@@ -5,7 +5,7 @@ type Treasury struct {
 	AddrId    int64  `gorm:"column:addr_id"` // stake_address(id)
 	CertIndex int32  `gorm:"column:cert_index"`
 	Amount    uint64 `gorm:"column:amount"` // This is a "int65type" column
-	TxId      int64  `gorm:"column:tx_id"`    // tx(id)
+	TxId      int64  `gorm:"column:tx_id"`  // tx(id)
 }
 
 // Override default table name
