+++ internal/datasource/cardano_db_sync/models/tx_metadata.go
@@ -1,11 +1,11 @@
 package models
 
 type TxMetadata struct {
-	Id             int64  `gorm:"column:id"`
-	Key            string `gorm:"column:key"` // This is a "word64type" column
-	Json           jsonb  `gorm:"column:json"`  // This is a "jsonb" type
-	Bytes          []byte `gorm:"column:bytes"`
-	TxId           int64  `gorm:"column:tx_id"` // tx(id)
+	Id    int64  `gorm:"column:id"`
+	Key   string `gorm:"column:key"`  // This is a "word64type" column
+	Json  jsonb  `gorm:"column:json"` // This is a "jsonb" type
+	Bytes []byte `gorm:"column:bytes"`
+	TxId  int64  `gorm:"column:tx_id"` // tx(id)
 }
 
 // Override default pluralized table name
